refactor: Reorganize initialization logic to improve testability
