+++ src/index.html
@@ -10,7 +10,7 @@
 	</head>
 	<body class="gutenberg-kit wp-embed-responsive">
 		<div id="root" class="gutenberg-kit-root"></div>
-		<script type="module" src="/index.jsx"></script>
+		<script type="module" src="/index.js"></script>
 		<div
 			id="popover-fallback-container"
 			class="components-popover__fallback-container"
