+++ src/index.js
@@ -1,0 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import { initializeBundledEditor } from './utils/bundled-editor';
+import './index.scss';
+
+initializeBundledEditor();
