+++ src/remote.js
@@ -1,0 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import { initializeRemoteEditor } from './utils/remote-editor';
+import './index.scss';
+
+initializeRemoteEditor();
